+++ CHANGELOG.md
@@ -10,6 +10,20 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Added `Clone()` method to all collection types.
+
+### Changed
+
+- **[BC]** Renamed `collection.UnorderedSet` to `Set`.
+
+### Removed
+
+- **[BC]** Removed `collection.Clone()`.
+
 ## [0.14.0] - 2024-09-30
 
 ### Added